MINOR: debug: report the time since last wakeup and call
When task profiling is enabled, the current thread knows when the
currently running task was woken up and called, so we can calculate
how long ago it was woken up and called. This is convenient to figure
whether or not a warning or panic is caused by this task or by a
previous one, so let's report this info in thread outputs when known.

It would be useful to backport this to 3.2.

(cherry picked from commit 25195ba1e77df8297b3361b275f5e4430e2729d8)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>