MINOR: debug: report the number of loops and ctxsw for each thread
When multiple similar warnings are emitted, it can be difficult to know
whether only one task is looping slowly or if many are sharing the CPU.
Let's report the number of context switches and polling loop turns in
thread dumps so that warnings are easier to understand.

This should be backported to 3.2.

(cherry picked from commit 12bc4f9c44beaa1a73b16171f63182f2b8140e3b)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>