MINOR: activity: indicate the number of calls on "show tasks"
The "show tasks" command can be useful to inspect run queues for active
tasks, but currently it's difficult to distinguish an occasional running
task from a heavily active one. Let's collect the number of calls for
each of them, report them average on the number of instances of each task
as well as a percentage of the total used. This way it even becomes
possible to get a hint about how CPU usage is distributed.

(cherry picked from commit 9eb35563a654d24544dd037aeef0494793070c7e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>