MEDIUM: resolvers: make the process_resolvers() task single-threaded
This task is sometimes caught triggering the watchdog while waiting for
the infamous resolvers lock, or the scheduler's wait queue lock in
task_queue(). Both are caused by its multi-threaded capability. The
task may indeed start on a thread that's different from the one that
is currently receiving a response and that holds the resolvers lock,
and when being queued back, it requires to lock the wait queue. Both
problems disappear when sticking it to a single thread. But for configs
running multiple resolvers sections, it would be suboptimal to run them
all on the same thread. In order to avoid this, we implement a counter
in the resolvers_finalize_config() section that rotates the thread for
each resolvers section.

This was sufficient to further improve the performance here, making the
CPU usage drop to about 7% (from 11 previously or 38 initially) and not
showing any resolvers lock contention anymore in perf top output.

The change was kept fairly minimal to permit a backport once enough
testing is conducted on it. It could address a significant part of
the trouble reported by Felipe in GH issue #3101.

(cherry picked from commit 2ce5e0edcce03d648a443c1f84307e3bdc7a2710)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 86a26a262aba4fb651d37aed7332f560fc07945d)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 24c7aa9db08cd57329f2807502bd4834393fd20f)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>