BUG/MEDIUM: resolvers: Properly cache do-resolv resolution
As stated by the documentation, when a do-resolv resolution is performed,
the result should be cached for <hold.valid> milliseconds. However, the only
way to cache the result is to always have a requester. When the last
requester is unlink from the resolution, the resolution is released. So, for
a do-resolv resolution, it means it could only work by chance if the same
FQDN is requested enough to always have at least two streams waiting for the
resolution. And because in that case, the cached result is used, it means
the traffic must be quite high.

In fact, a good approach to fix the issue is to keep orphan resolutions to
be able cache the result and only release them after hold.valid milliseconds
after the last real resolution. The resolver's task already releases orphan
resolutions. So we only need to check the expiration date and take care to
not release the resolution when the last stream is unlink from it.

This patch should be backported to all stable versions. We can start to
backport it as far as 3.1 and then wait a bit.

(cherry picked from commit 37abe56b186356f570d63d0f6e2fb4b851ae8660)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 274cc7837a4012ba252e91b1b3bd3272c021363d)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit a2d2a6fe3c0980db54b2f6856193c0c8db2c4d6b)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 9fcada8e95924395997e723e64562f931771b7a8)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>