MINOR: backend: srv_is_up converter
There is currently an srv_queue converter which is capable of taking the
output of a dynamic name and determining the queue length for a given
server.  In addition there is a sample fetcher for whether a server is
currently up.  This simply combines the two such that srv_is_up can be
used as a converter too.

Future work might extend this to other sample fetchers for servers, but
this is probably the most useful for acl routing.

(cherry picked from commit 54f53bc8756af503e6ca521a95d85dbeded1aedf)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>