BUG/MEDIUM: acme: free() of i2d_X509_REQ() with AWS-LC
When using AWS-LC, the free() of the data ptr resulting from
i2d_X509_REQ() might crash, because it uses the free() of the libc
instead of OPENSSL_free().

It does not seems to be a problem on openssl builds.

Must be backported in 3.2.

(cherry picked from commit 3ce597bfa23d6e4b8ec62ce22c65b677c328027b)
Signed-off-by: William Lallemand <wlallemand@haproxy.com>